model retrained, added confidence score
