+++ cone_lcm_publisher.py
@@ -18,13 +18,13 @@ if __name__ == '__main__':
     calibration_data = np.load('cam_calibration_data.npz')
 
     # Load a YOLO PyTorch model
-    # model = YOLO("utils/cone_detection_model.pt")
+    # model = YOLO("utils/best.pt")
 
     # Export the model to NCNN format
     # model.export(format="ncnn")
 
     # Load the exported NCNN model
-    ncnn_model = YOLO("utils/cone_detection_model_ncnn_model")
+    ncnn_model = YOLO("utils/best_ncnn_model", task='detect')
 
     camera = Camera(camera_id, image_width, image_height, fps)
     register_signal_handlers(camera.cleanup)
