+++ utils/best_ncnn_model/metadata.yaml
@@ -1,0 +1,18 @@
+description: Ultralytics YOLO11s model trained on /content/MBot-Cone-Detection-5/data.yaml
+author: Ultralytics
+date: '2024-11-12T19:35:33.303260'
+version: 8.3.29
+license: AGPL-3.0 License (https://ultralytics.com/license)
+docs: https://docs.ultralytics.com
+stride: 32
+task: detect
+batch: 1
+imgsz:
+- 640
+- 640
+names:
+  0: blue_cone
+  1: green_cone
+  2: pink_cone
+  3: red_cone
+  4: yellow_cone
